# Backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and generate passport token",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User's password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Successful logout"
                    },
                    "400": {
                        "description": "Invalid token"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get logged-in user details",
                "operationId": "68f71ba285837ee66901e2681a219252",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "User's id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "parameters": {
            "name": {
                "name": "name",
                "in": "query",
                "description": "User's name",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "email": {
                "name": "email",
                "in": "query",
                "description": "User's email",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "password": {
                "name": "password",
                "in": "query",
                "description": "User's password",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "user_id": {
                "name": "user_id",
                "in": "query",
                "description": "User's id",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "1": {
                "name": "user_id",
                "in": "query",
                "description": "User's id",
                "required": true,
                "schema": {
                    "type": "string"
                }
            },
            "role": {
                "name": "role",
                "in": "query",
                "description": "Role's name",
                "required": true,
                "schema": {
                    "type": "string"
                }
            }
        }
    }
}